requirements.txt
Django==1.8.2
djangorestframework==3.1.3
flake8==2.4.1
Markdown==2.6.2
mccabe==0.3.1
pep8==1.5.7
pyflakes==0.8.1
uWSGI==2.0.10
wheel==0.24.0
